Refactor StockList
