+++ lib/inventree/stock.dart
@@ -149,7 +149,8 @@ class InvenTreeStockItem extends InvenTreeModel {
       "part_detail": "true",
       "location_detail": "true",
       "supplier_detail": "true",
-      "cascade": "false"
+      "cascade": "false",
+      "in_stock": "true",
     };
   }
